+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- adds whoami functionality that will show the current logged in user (`deploy whoami`)
+
 # 0.0.1 (08/08/2017)
 
 - basic functionality working